Refactored form into diff components, set prop-types to off
